--- src/server/localisations/infra/repositories/apiAdresse.mapper.ts
@@ -1,3 +1,0 @@
-export function removeParenthesis(query: string): string { // FIXME: Meilleur nom de fichier ?
-	return query.replace(/[()]/g, '');
-}
