+++ src/server/localisations/infra/repositories/apiAdresse.repository.ts
@@ -3,7 +3,7 @@ import { RésultatsRechercheCommune } from '~/server/localisations/domain/locali
 import {
 	LocalisationAvecCoordonnéesRepository,
 } from '~/server/localisations/domain/localisationAvecCoordonnées.repository';
-import { removeParenthesis } from '~/server/localisations/infra/repositories/apiAdresse.mapper';
+import { removeParenthesis } from '~/server/localisations/infra/repositories/removeParenthesis';
 import { ApiAdresseResponse } from '~/server/localisations/infra/repositories/apiAdresse.response';
 import { mapRésultatsRechercheCommune } from '~/server/localisations/infra/repositories/apiGeo.mapper';
 import { ErrorManagementService } from '~/server/services/error/errorManagement.service';
