+++ src/server/localisations/infra/repositories/removeParenthesis.ts
@@ -1,0 +1,3 @@
+export function removeParenthesis(query: string): string {
+	return query.replace(/[()]/g, '');
+}
